Reduce overhead of activity launch observer
The cost of writing ActivityRecordProto is heavy. Especially
it executes in window manager's lock.

Since the usage of proto was removed. The callback interface
is simplified to only send id and name. Though currently there
is no use case, it is more accurate to associate different launch
sequence because the assumption "only one start at a time" is not
true anymore since multi-window/area/display support.

Bug: 229822111
Test: atest ActivityMetricsLaunchObserverTests

Change-Id: I6a151e72616fd4cf7a2e17d3c7f92aee8605c485