Filter instant app from standby bucket
Since we restrict access to instant app data in usage stats (by
replacing instant app's package name with a placeholder name), we
should also filter out this data when getting standby bucket with
a given package name. This gives the callers a way to tell if the
instant app exists or not.

Bug: 217561828
Bug: 217740016
Test: atest CtsUsageStatsTestCases
Test: manually using the PoC in the buganizer to ensure the symptom
      no longer exists.
Change-Id: I091e65c01d184b9df02fda29c64ac61787077fdb