Don't show the bouncer onPanelExpansionChanged when dozing + not pulsing
When the device goes to sleep from the power button,
the call to show the keyguard on AOD updates the
NotificationPanel expansion to show to see the AOD views.

However, with certain timing (keyguardShowing=true 'too early'),
the panel expansion change ends up showing the bouncer since
the panel expansion controls the bouncer when keyguardShowing=true.

Showing the bouncer subsequently wakes up the device,
resulting in the device waking up right after the user
pressed the power button with the intention to enter AoD.

This CL prevents onPanelExpansionChanged from effecting the
bouncer expansion when the device is dozing (and not pulsing).

Test: press power button from home screen & when qs is expanded,
notice that device consistently goes to sleep and doesn't immediately
wake back up
Test: fail udfps twice on AoD, see bouncer still apears
Test: receive a notification in AoD w/ an intent, tap on notification
and see bouncer (UDFPS or primary auth) appears
Fixes: 228788448

Change-Id: Ifb8888900a3bd2f229f4bfdda9a59f0778efc5dd