Ensure light reveal scrim is fully revealed once we're unlocked.
We've had some difficult to reproduce bugs where the
lightreveal remains visible in the shade while
unlocked. These are likely related to bugs in the
keyguard state code, which we intend to refactor in
U.

In the meantime, the scrim should never be anything
but fully revealed once the keyguard going away
transition ends.

Fixes: 229778554
Test: manually force all lightreveal animations to end at 0.5f, then unlock
Test: can't add test to centralsurfacestest because start() is not called
Change-Id: I6bd65738635a2f966227014a4920f8e8367b91e6