Delete config_killableInputMethods setting
This change effectively reverts the introduction of the
config_killableInputMethods setting in [1]. Permanently setting the
oom_score_adj of the IME process to CACHED leads to cases where the
lowmemorykiller kills it before it can be raised to "visible". In
order to achieve the killable IME behavior, enabling the setting
config_preventImeStartupUnlessTextEditor is sufficient.

Test: atest CtsInputMethodTestCases
Test: also manually on adt3
Bug: 232511947
[1]: Ib191d964de15bbb84daa8bbeca6f2dca6f96311c

Change-Id: I92091e0e1c26635b8fc38f1371445a17a0afafae