Demote top app scheduling group when recents is animating
With legacy transition, the recents activity doesn't move to top
when it is animating. So both the real top app and the recents
process will have top-app scheduling group. Then there may be
a CPU contention which causes animation jank.

Because the recents animation is usually more noticeable (mainly
swipe-up to home), demote the current top app to foreground
group for reducing the potential contention.

Bug: 233992724
Test: atest RecentsAnimationTest
Test: adb shell dumpsys activity o | grep -e "A.*calc"
  Launch calculator:
    T/A/TOP  LCMN  t: 0 22249:calc/u0a158 (top-activity)
  Swipe-up to enter recents:
    F/A/TOP  LCMN  t: 0 22249:calc/u0a158 (intermediate-top-activity)
  Return to calculator:
    T/A/TOP  LCMN  t: 0 22249:calc/u0a158 (top-activity)
  Swipe to home:
    b/ /LAST ----  t: 0 22249:calc/u0a158 (previous)

Change-Id: I0d002ea46de4474fcf863f142f2ee17fe3a5c742