Clean up QS/NSSL bottom margin calculation
The root cause of b/222594381 was a deviation between the two mostly
repetitive branches of an if-else.

This change factors out the not-repetitive part of those two branches
into a top-level case, and then merges the rest of the cases between the
two branches.

It also removes the default value for stackScrollMargin, since that
saves a couple lines of code but makes the resulting outputs harder to
discern by eye.

Also, spell out some wildcard imports to make ktlint happy.

Bug: 222594381
Test: manual, on Pixel 5a
Change-Id: I6a034a6fb475be363667cb64ab2b3516a8df6005